+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Update to rubocop 0.37.1. ([@nijikon][])
 * Update ruby versions we test against. ([@nijikon][])
+* Add `RSpec::NotToNot` cop. ([@miguelfteixeira][])
 
 ## 1.3.1
 
@@ -71,3 +72,4 @@
 [@nevir]: https://github.com/nevir
 [@nijikon]: https://github.com/nijikon
 [@pstengel]: https://github.com/pstengel
+[@miguelfteixeira]: https://github.com/miguelfteixeira